ads-diagnostics-platform/requirements.txt
fastapi
uvicorn
streamlit
pandas
numpy
requests

sentence-transformers
faiss-cpu
transformers
torch

scikit-learn